Changed the u64 type usage in usize such that x86 targets work as well
Tested with the i686-unknown-linux-gnu target and works just as
expected. Also, inline assembly for x86 intel had to be changed as well
(IP-relative addressing requires 64-bit mode).

+++ src/address.rs
@@ -1,7 +1,7 @@
 use std::process;
 use std::fs;
 
-pub fn get_code_address(_binary_name: &str) -> u64 {
+pub fn get_code_address(_binary_name: &str) -> usize {
     let pid = process::id();
     let proc_maps_filename = format!("/proc/{}/maps", pid);
 
@@ -9,5 +9,5 @@ pub fn get_code_address(_binary_name: &str) -> u64 {
 
     let code_address_hex = contents.split("-").next().unwrap();
 
-    u64::from_str_radix(code_address_hex, 16).unwrap()
+    usize::from_str_radix(code_address_hex, 16).unwrap()
 }
